Fixed typings

+++ test/unit/index.js
@@ -36,7 +36,7 @@ Test('create table', () => {
 
 	let table = database.addTable('cars');
 
-	let cars = database.find('cars');
+	let cars = database.select('cars');
 
 	assert.deepStrictEqual(cars.length, 0);
 });
@@ -46,10 +46,11 @@ Test('create row', () => {
 
 	let table = database.addTable('cars');
 	let car = database.create('cars', { color: 'blue' });
+	if (car == undefined) throw new Error(`Should have created a car`);
 
 	assert.deepStrictEqual(car.color, 'blue');
 
-	let cars = database.find('cars');
+	let cars = database.select('cars');
 
 	assert.deepStrictEqual(cars.length, 1);
 	assert.deepStrictEqual(cars[0].color, 'blue');
@@ -62,9 +63,9 @@ Test('create row with id', () => {
 
 	let car = database.create('cars', { id: 'abc', color: 'blue' });
 
-	assert.deepStrictEqual(car.id, 'abc');
+	assert.deepStrictEqual(car?.id, 'abc');
 
-	let cars = database.find('cars');
+	let cars = database.select('cars');
 
 	assert.deepStrictEqual(cars[0].id, 'abc');
 });
@@ -75,10 +76,10 @@ Test('create row without id', () => {
 	let table = database.addTable('cars');
 
 	let car = database.create('cars', { color: 'blue' });
-	let cars = database.find('cars');
+	let cars = database.select('cars');
 
-	assert.deepStrictEqual(car.id, cars[0].id);
-	assert.notDeepStrictEqual(car.id, undefined);
+	assert.deepStrictEqual(car?.id, cars[0].id);
+	assert.notDeepStrictEqual(car?.id, undefined);
 });
 
 Test('create row without table', () => {
@@ -134,7 +135,7 @@ Test('create row before index', () => {
 	let car2 = database.create('cars', { color: 'blue' });
 	let car3 = database.create('cars', { color: 'red' });
 	let index = database.addIndex('cars', 'color');
-	let cars = database.find('cars', { color: 'red' });
+	let cars = database.select('cars', { color: 'red' });
 
 	assert.deepStrictEqual(cars.length, 2);
 	assert.deepStrictEqual(cars[0].color, 'red');
@@ -149,7 +150,7 @@ Test('create row after index', () => {
 	let car1 = database.create('cars', { color: 'red' });
 	let car2 = database.create('cars', { color: 'blue' });
 	let car3 = database.create('cars', { color: 'red' });
-	let cars = database.find('cars', { color: 'red' });
+	let cars = database.select('cars', { color: 'red' });
 
 	assert.deepStrictEqual(cars.length, 2);
 	assert.deepStrictEqual(cars[0].color, 'red');
@@ -165,7 +166,7 @@ Test('create index with multiple attributes', () => {
 	let car2 = database.create('cars', { brand: 'chevrolet', color: 'blue' });
 	let car3 = database.create('cars', { brand: 'chevrolet', color: 'red' });
 	let car4 = database.create('cars', { brand: 'chevrolet', color: 'red' });
-	let cars = database.find('cars', { brand: 'chevrolet', color: 'red' });
+	let cars = database.select('cars', { brand: 'chevrolet', color: 'red' });
 
 	assert.deepStrictEqual(cars.length, 2);
 	assert.deepStrictEqual(cars[0].color, 'red');
@@ -179,11 +180,12 @@ Test('delete row', () => {
 
 	database.addTable('cars');
 	let car = database.create('cars', { color: 'red' });
+	if (car == undefined) throw new Error(`Should have created a car`);
 	let deletedCar = database.delete('cars', car);
-	let cars = database.find('cars');
+	let cars = database.select('cars');
 
 	assert.deepStrictEqual(cars.length, 0);
-	assert.deepStrictEqual(deletedCar.color, 'red');
+	assert.deepStrictEqual(deletedCar?.color, 'red');
 });
 
 Test('delete row as id', () => {
@@ -192,10 +194,10 @@ Test('delete row as id', () => {
 	let table = database.addTable('cars');
 	let car = database.create('cars', { id: '123', color: 'red' });
 	let deletedCar = database.delete('cars', '123');
-	let cars = database.find('cars');
+	let cars = database.select('cars');
 
 	assert.deepStrictEqual(cars.length, 0);
-	assert.deepStrictEqual(deletedCar.color, 'red');
+	assert.deepStrictEqual(deletedCar?.color, 'red');
 });
 
 Test('delete row with id', () => {
@@ -204,10 +206,10 @@ Test('delete row with id', () => {
 	let table = database.addTable('cars');
 	let car = database.create('cars', { id: '123', color: 'red' });
 	let deletedCar = database.delete('cars', { id: '123' });
-	let cars = database.find('cars');
+	let cars = database.select('cars');
 
 	assert.deepStrictEqual(cars.length, 0);
-	assert.deepStrictEqual(deletedCar.color, 'red');
+	assert.deepStrictEqual(deletedCar?.color, 'red');
 });
 
 Test('delete table', () => {
@@ -219,7 +221,7 @@ Test('delete table', () => {
 	database.removeTable('cars');
 	database.addTable('cars');
 
-	let cars = database.find('cars');
+	let cars = database.select('cars');
 
 	assert.deepStrictEqual(cars.length, 0);
 });
@@ -229,9 +231,10 @@ Test('update', () => {
 
 	database.addTable('cars');
 	let oldCar = database.create('cars', { brand: 'Tesla', color: 'red' });
+	if (oldCar == undefined) throw new Error('Should have created a car');
 	let newCar = database.update('cars', oldCar, { color: 'blue' });
 
-	let cars = database.find('cars');
+	let cars = database.select('cars');
 
 	assert.deepStrictEqual(cars.length, 1);
 	assert.deepStrictEqual(cars[0].brand, 'Tesla');
@@ -245,7 +248,7 @@ Test('update by id', () => {
 	let oldCar = database.create('cars', { id: 123, brand: 'Tesla', color: 'red' });
 	let newCar = database.update('cars', 123, { color: 'blue' });
 
-	let cars = database.find('cars');
+	let cars = database.select('cars');
 
 	assert.deepStrictEqual(cars.length, 1);
 	assert.deepStrictEqual(cars[0].brand, 'Tesla');
@@ -257,9 +260,10 @@ Test('replace', () => {
 
 	database.addTable('cars');
 	let oldCar = database.create('cars', { brand: 'Tesla', color: 'red' });
+	if (oldCar == undefined) throw new Error('Should have created a car');
 	let newCar = database.replace('cars', oldCar, { color: 'blue' });
 
-	let cars = database.find('cars');
+	let cars = database.select('cars');
 
 	assert.deepStrictEqual(cars.length, 1);
 	assert.deepStrictEqual(cars[0].brand, undefined);
@@ -273,7 +277,7 @@ Test('replace by id', () => {
 	let oldCar = database.create('cars', { id: 123, brand: 'Tesla', color: 'red' });
 	let newCar = database.replace('cars', 123, { color: 'blue' });
 
-	let cars = database.find('cars');
+	let cars = database.select('cars');
 
 	assert.deepStrictEqual(cars.length, 1);
 	assert.deepStrictEqual(cars[0].brand, undefined);
@@ -310,9 +314,12 @@ Test('dynamic functions', () => {
 	let database = new Database(localStorage);
 
 	let table = database.addTable('cars', 'car');
+	// @ts-ignore
 	let car1 = database.createCar({ color: 'red' });
+	// @ts-ignore
 	let car2 = database.createCar({ color: 'blue' });
-	let cars = database.findCars();
+	// @ts-ignore
+	let cars = database.selectCars();
 
 	assert.deepStrictEqual(cars.length, 2);
 	assert.deepStrictEqual(cars[0].color, 'red');
@@ -323,8 +330,9 @@ Test('select with id property', () => {
 	let database = new Database(localStorage);
 
 	let table = database.addTable('cars', 'car');
-	let car = database.createCar({ color: 'red' });
-	let cars = database.find('cars', { id: car.id });
+	let car = database.create('cars', { color: 'red' });
+	if (car == undefined) throw new Error('Should have created a car');
+	let cars = database.select('cars', { id: car.id });
 
 	assert.strictEqual(cars.length, 1);
 	assert.strictEqual(cars[0].color, car.color);
@@ -334,8 +342,9 @@ Test('select with id property and autoindex', () => {
 	let database = new Database(localStorage);
 
 	let table = database.addTable('cars', 'car');
-	let car = database.createCar({ color: 'red' });
-	let cars = database.find('cars', { id: car.id }, true);
+	let car = database.create('cars', { color: 'red' });
+	if (car == undefined) throw new Error('Should have created a car');
+	let cars = database.select('cars', { id: car.id }, true);
 
 	assert.strictEqual(cars.length, 1);
 	assert.strictEqual(cars[0].color, car.color);
@@ -345,8 +354,9 @@ Test('select with id and equal properties', () => {
 	let database = new Database(localStorage);
 
 	let table = database.addTable('cars', 'car');
-	let car = database.createCar({ color: 'red' });
-	let cars = database.find('cars', { id: car.id, color: 'red' }, true);
+	let car = database.create('cars', { color: 'red' });
+	if (car == undefined) throw new Error('Should have created a car');
+	let cars = database.select('cars', { id: car.id, color: 'red' }, true);
 
 	assert.strictEqual(cars.length, 1);
 	assert.strictEqual(cars[0].color, car.color);
@@ -356,8 +366,110 @@ Test('select with id and different properties', () => {
 	let database = new Database(localStorage);
 
 	let table = database.addTable('cars', 'car');
-	let car = database.createCar({ color: 'red' });
-	let cars = database.find('cars', { id: car.id, color: 'blue' }, true);
+	let car = database.create('cars', { color: 'red' });
+	if (car == undefined) throw new Error('Should have created a car');
+	let cars = database.select('cars', { id: car.id, color: 'blue' }, true);
 
 	assert.strictEqual(cars.length, 0);
 });
+
+Test('select by id', () => {
+	let database = new Database(localStorage);
+
+	let table = database.addTable('cars', 'car');
+	let car1 = database.create('cars', { id: 'A', color: 'red' });
+	if (car1 == undefined) throw new Error('Should have created a car');
+	let car2 = database.create('cars', { id: 'B', color: 'blue' });
+	if (car2 == undefined) throw new Error('Should have created a car');
+
+	let car = database.selectById('cars', 'B');
+
+	assert.deepStrictEqual(car, car2);
+});
+
+Test('select ids', () => {
+	let database = new Database(localStorage);
+
+	let table = database.addTable('cars', 'car');
+	let car1 = database.create('cars', { id: 'A', color: 'red' });
+	if (car1 == undefined) throw new Error('Should have created a car');
+	let car2 = database.create('cars', { id: 'B', color: 'blue' });
+	if (car2 == undefined) throw new Error('Should have created a car');
+
+	let carIds = database.selectIds('cars');
+
+	assert.deepStrictEqual(carIds, ['A', 'B']);
+});
+
+Test('select ids with properties', () => {
+	let database = new Database(localStorage);
+
+	let table = database.addTable('cars', 'car');
+	let car1 = database.create('cars', { id: 'A', color: 'red' });
+	if (car1 == undefined) throw new Error('Should have created a car');
+	let car2 = database.create('cars', { id: 'B', color: 'blue' });
+	if (car2 == undefined) throw new Error('Should have created a car');
+
+	let carIds = database.selectIds('cars', { color: 'blue' }, true);
+
+	database.select('cars');
+
+	assert.deepStrictEqual(carIds, ['B']);
+});
+
+Test.only('huh?', () => {
+	let database = new Database(localStorage);
+
+	database.addTable('todos', 'todo');
+	database.addTable('settings', 'setting');
+
+	database.create('settings', { id: 'introduction', value: false });
+
+	for (let key in localStorage) {
+		console.log(key, localStorage.getItem(key));
+	}
+});
+
+// Test('relations', () => {
+// 	let database = new Database(localStorage)
+
+// 	let carsTable = database.addTable('cars', 'car');
+// 	let tripsTable = database.addTable('trips', 'trip')
+// 	let relation = database.addRelation('cars', 'trips')
+
+// 	database.selectById('cars', 'A', { includes: 'trips' })
+// 	database.select('cars', )
+// })
+
+// database.addRelation('game', 'player', optionalAttributeName = arg2)
+// database.addRelation('game', 'players')
+// database.addRelation('games', 'players')
+// database.addRelation('games', 'player')
+
+// Test('select with array properties', () => {
+// 	let database = new Database(localStorage);
+// 	let table = database.addTable('cars', 'car');
+// 	let car1 = database.create('cars', { id: 'A', color: 'red' });
+// 	if (car1 == undefined) throw new Error('Should have created a car');
+// 	let car2 = database.create('cars', { id: 'B', color: 'blue' });
+// 	if (car2 == undefined) throw new Error('Should have created a car');
+// 	let car3 = database.create('cars', { id: 'C', color: 'green' });
+// 	if (car3 == undefined) throw new Error('Should have created a car');
+// 	let carIds = database.selectIds('cars', { color: ['blue', 'green', 'pink'] }, true);
+// 	assert.deepStrictEqual(carIds, ['B']);
+// });
+
+// Test('relations', () => {
+// 	let database = new Database(localStorage);
+// 	let carsTable = database.addTable('cars', 'car');
+// 	let ownerTable = database.addTable('owners', 'owner');
+// 	database.addRelation('car.owner', 'owner');
+// 	database.addRelation('car.owners', 'owners');
+// 	database.select('cars');
+// 	database.
+// });
+
+// Database.addRelation('game.player', 'player');
+// Database.addRelation('games.player', 'player');
+// Database.addRelation('game.player', 'players');
+// Database.addRelation('games.players', 'players.games');
